+++ ccramic/inputs/pixel_level_inputs.py
@@ -312,15 +312,19 @@ def update_canvas_filename(canvas_config: dict, roi_name: str=None, delimiter: s
 
 def set_canvas_viewport(size_slider_val: Union[float, int]=None,
                         image_dict: dict=None, data_selection: str=None,
-                        current_canvas: Union[go.Figure, dict]=None, cur_canvas_layout: dict=None):
+                        current_canvas: Union[go.Figure, dict]=None, cur_canvas_layout: dict=None,
+                        cur_dimensions: Union[tuple, list]=None):
     """
     Set the canvas viewport based on the canvas size range slider, as well as the aspect ratio of
     the ROI dimensions
     returns a hash for the width and height in vh: {'width': f'{value}vh', 'height': f'{value}vh'}
     """
     try:
-        first_image = get_first_image_from_roi_dictionary(image_dict[data_selection])
-        aspect_ratio = int(first_image.shape[1]) / int(first_image.shape[0])
+        if cur_dimensions is None:
+            first_image = get_first_image_from_roi_dictionary(image_dict[data_selection])
+            aspect_ratio = int(first_image.shape[1]) / int(first_image.shape[0])
+        else:
+            aspect_ratio = int(cur_dimensions[1]) / int(cur_dimensions[0])
     except (KeyError, AttributeError, IndexError):
         if current_canvas is not None and 'layout' in current_canvas and \
                 'range' in current_canvas['layout']['xaxis'] and \